third_party/blink/web_tests/flag-specific/enable-skia-graphite/external/wpt/mediacapture-record/MediaRecorder-canvas-media-source.https-expected.txt

This is a testharness.js-based test.
[FAIL] MediaRecorder returns frames containing video content
  assert_greater_than: expected a number greater than 1500 but got 1
Harness: the test ran to completion.
